Action관련한 기본적 ui기능 구현 통계 작업 구현중

+++ data/DataRoom/src/main/java/etude/de/lapin/baby/data/room/model/ActionEntity.kt
@@ -13,7 +13,7 @@ data class ActionEntity (
     val categoryId: Int,
 
     @ColumnInfo(name = "volume")
-    val volume: Int? = null,
+    val volume: Float? = null,
 
     @ColumnInfo(name = "timestamp")
     val timestamp: Long,
