+++ data/DataRoom/src/main/java/etude/de/lapin/baby/data/room/model/ActionResultEntity.kt
@@ -4,7 +4,7 @@ data class ActionResultEntity (
     val id: Int = 0,
     val categoryId: Int,
     val categoryName: String,
-    val volume: Int? = null,
+    val volume: Float? = null,
     val timestamp: Long,
     val memo: String? = null
